Rework the variable tree to be typed and have descriptions
The previous incantation of the VarTree did not provide information
nor constraints on the variables that it can hold.  This made it
difficult to provide information about the structure of the tree to
the interfaces etc.

Furthermore, each of the types that lives in the tree has a limited
API that will be made available through each of the interfaces.

This patch changes the tree to have the following variable types:

 - Integers
 - Strings
 - Lists
 - Dictionaries

Dictionaries are like Python's dictionary type.  They are a
name-to-value map, and can contain any of the variable types
are listed above, including themselves.